requirements.txt
matplotlib==3.7.2
numpy==1.24.2
networkx==3.0
pandas==1.5.3
torch==1.13.1
scikit-learn==1.2.1
scipy==1.10.1
julia==0.6.1
hydra-core==1.3.1
kaleido==0.2.1
plotly==5.14.1
